Use Floats again
Not using floats breaks everything. We must wait a little longer before we
can enact the Integer Revolution!

+++ src/CrossyToad/Geometry/Offset.hs
@@ -9,4 +9,4 @@ import Linear.V2
 -- | to have any meaningful location
 -- |
 -- | In reality it's just a handy name for one use of a vector
-type Offset = (V2 Int)
+type Offset = (V2 Float)
